Split module `Bech32` into public API and private `Internal` module.
